third_party/blink/web_tests/platform/linux-chrome/external/wpt/window-management/multi-screen-window-open-fullscreen.tentative.https-expected.txt

This is a testharness.js-based test.
[PASS] Use multi-screen details to open a fullscreen popup window on each screen
[FAIL] Open a fullscreen popup on '0'
  assert_equals: popup: (250,250 300x100), 'undefined': (undefined,undefined 800x600) expected 0 but got 250
Harness: the test ran to completion.
